# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Spending Decision Support API
  version: 1.2.0
  description: >-
    Versioned API for personal spending forecasts, unusual-spending alerts,
    and transparent recommendations. Unusual-spending results are not fraud
    classifications, and recommendations are not professional financial advice.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a user
      security: []
      x-request-example:
        email: user@example.com
        password: StrongPass123!
        display_name: Eva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: StrongPass123!
              display_name: Eva
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                success: true
                data:
                  access_token: eyJ...
                  token_type: Bearer
                  expires_in: 3600
                  user:
                    id: 5ce4cc8a-0bbe-4d02-a1c4-483cc215d476
                    email: user@example.com
                    display_name: Eva
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      security: []
      x-request-example:
        email: user@example.com
        password: StrongPass123!
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: StrongPass123!
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                success: true
                data:
                  access_token: eyJ...
                  token_type: Bearer
                  expires_in: 3600
                  user:
                    id: 5ce4cc8a-0bbe-4d02-a1c4-483cc215d476
                    email: user@example.com
                    display_name: Eva
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Return the authenticated user
      x-request-example: {}
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                success: true
                data:
                  id: 5ce4cc8a-0bbe-4d02-a1c4-483cc215d476
                  email: user@example.com
                  display_name: Eva
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/google:
    post:
      tags: [Authentication]
      summary: Exchange a verified Google ID token for a Spendly session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token: {type: string, minLength: 20, maxLength: 8192}
      responses:
        '200':
          description: Authenticated or safely linked
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
  /auth/profile:
    put:
      tags: [Authentication]
      summary: Change the authenticated user's username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[A-Za-z0-9_]+$'
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /auth/onboarding:
    put:
      tags: [Authentication]
      summary: Persist completion or skipping of new-user onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: {type: string, enum: [completed, skipped]}
      responses:
        '200':
          description: Onboarding state saved
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: A valid access token is required.
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: One or more request fields are invalid.
              details:
                amount: Amount must be greater than zero.
...
    Conflict:
      description: Resource conflicts with existing data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: EMAIL_ALREADY_REGISTERED
              message: An account with this email already exists.
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            details:
              type: object
              additionalProperties: true
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: {type: string, format: email, maxLength: 255}
        password: {type: string, minLength: 10, maxLength: 128}
        display_name: {type: string, minLength: 1, maxLength: 100}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}
    User:
      type: object
      required: [id, email, display_name, username, onboarding_completed]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        display_name: {type: string}
        username: {type: string, minLength: 3, maxLength: 30}
        onboarding_completed: {type: boolean}
        has_required_consents: {type: boolean}
    AuthResponse:
      type: object
      properties:
        success: {type: boolean}
        data:
          type: object
          properties:
            access_token: {type: string}
            token_type: {type: string}
            expires_in: {type: integer}
            user: {$ref: '#/components/schemas/User'}
    UserResponse:
      type: object
      properties:
        success: {type: boolean}
        data: {$ref: '#/components/schemas/User'}
...